#192: remove dead ActivityProvider seam; dedup display-name, hidden-file filter, NoOp picker
