+++ composeApp/src/androidMain/kotlin/com/tubetoast/tether/TetherApp.kt
@@ -18,7 +18,6 @@ class TetherApp :
         super.onCreate()
         val debuggable = (applicationInfo.flags and ApplicationInfo.FLAG_DEBUGGABLE) != 0
         initTetherLogging(debugEnabled = debuggable)
-        registerActivityLifecycleCallbacks(container.activityProvider)
         container.autoSendDispatcher.start()
     }
 }
