+++ composeApp/src/androidMain/kotlin/com/tubetoast/tether/di/AndroidAppContainer.kt
@@ -59,8 +59,6 @@ class AndroidAppContainer(
     )
     override val ownDeviceType: DeviceType = DeviceType.Android
 
-    val activityProvider: AndroidActivityProvider = AndroidActivityProvider()
-
     val pickerCoordinator: AndroidPickerCoordinator = AndroidPickerCoordinator()
 
     val androidFilePicker: AndroidFilePicker = AndroidFilePicker(
