+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/transfer/NoOpFilePicker.kt
@@ -1,0 +1,9 @@
+package com.tubetoast.tether.transfer
+
+internal object NoOpFilePicker : FilePicker {
+    override suspend fun pickFiles(): List<FileSource> = emptyList()
+
+    override suspend fun pickFolder(): List<FileSource> = emptyList()
+
+    override suspend fun pickPhotos(): List<FileSource> = emptyList()
+}
