--- composeApp/src/desktopMain/kotlin/com/tubetoast/tether/di/PlatformActivity.kt
@@ -1,3 +1,0 @@
-package com.tubetoast.tether.di
-
-actual class PlatformActivity
